Tutee is asking yes/no question when it is stuck
When tutee is stuck, currently it only asks tutor to demonstrate the step directly. This does not let tutor to convey that they might know a step. Therefore, tutee will now ask yes/no question if tutor knows what to do next or not?, if yes, it would let tutor demonstrate the step in the appropriate position as before. If no, that functionality should go in else part.