Commented out "Mr Williams flagged step why incorrect question" and "Both agree speech tell me more questions"
Before study 9, we had more initial tutee inquiry included in the CTI design. But based on data analysis, we found that followup helps and asking tutor to elaborate using conceptual terms help. So we decided to comment out the extra included ITIs and instead do a followup after the existing ITIs only which are "In past, I thought x is correct, why it does not apply now?" and "Why x?"